+++ tests/nouns/adjectives.test.js
@@ -13,5 +13,18 @@ test('.adjectives():', function(t) {
   doc = nlp('the cat who was really mean')
   m = doc.nouns().adjectives()
   t.equal(m.text(), 'mean', 'who was really .')
+
+  doc = nlp('the cat that was mean attacked the cute dog')
+  m = doc
+    .nouns(0)
+    .nouns()
+    .adjectives()
+  t.equal(m.text(), 'mean', 'first-noun')
+  m = doc
+    .nouns(1)
+    .nouns()
+    .adjectives()
+  t.equal(m.text(), 'cute', 'second-noun')
+
   t.end()
 })